# docs/openapi.yml
openapi: 3.0.3
info:
  title: Linkloop Pronunciation API
  version: 1.0.0
  description: |
    発音評価・進捗管理のためのAPI（Fastify実装）。
    
    - prefix: /api
    - 認証: x-user-idヘッダー（必要なAPIのみ）

servers:
  - url: http://localhost:3001/api

paths:
  /pronunciation-assessment:
    post:
      summary: 発音評価（基本）
      description: 音声データとテキストを送信し、Azureで発音評価を実施。結果を返す。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audio, referenceText]
              properties:
                audio:
                  type: string
                  format: binary
                  description: 音声データ（Base64エンコード）
                referenceText:
                  type: string
                  description: 評価対象のテキスト
      responses:
        '200':
          description: 発音評価結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  azureResponse:
                    type: object
                    description: Azure Speech ServiceのPronunciationAssessmentレスポンス
        '400':
          description: audio/referenceText未指定
        '500':
          description: サーバーエラー

  /pronunciation-assessment-advanced:
    post:
      summary: 発音評価（アドバンスト）
      description: 音声ファイルとテキストを送信し、Azureで発音評価を実施。結果を返す。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
                  description: 音声ファイル（wav/mp3等）
                referenceText:
                  type: string
                  description: 評価対象のテキスト
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
          required: false
          description: ユーザーID（集計時に必要）
      responses:
        '200':
          description: 発音評価結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  analysisId:
                    type: integer
                    description: 新しく作成された分析ID
                  azureResponse:
                    type: object
                    description: Azure Speech ServiceのPronunciationAssessmentレスポンス
        '400':
          description: audio/referenceText未指定
        '500':
          description: サーバーエラー

  /pronunciation-assessment/{analysisId}/feedback:
    get:
      summary: 発音評価のフィードバック取得
      description: 特定の分析IDに対するパーソナライズされたフィードバックを返す
      parameters:
        - in: path
          name: analysisId
          required: true
          schema:
            type: integer
          description: 分析ID
        - in: query
          name: focusOn
          schema:
            type: array
            items:
              type: string
          description: 重点的にフィードバックを取得したい音素の配列
        - in: query
          name: includeGeneralAdvice
          schema:
            type: boolean
          description: 一般的なアドバイスを含めるかどうか
      responses:
        '200':
          description: パーソナライズされたフィードバック
          content:
            application/json:
              schema:
                type: object
                properties:
                  sections:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                        content:
                          type: string
        '404':
          description: 分析が見つからない
        '500':
          description: サーバーエラー

  /pronunciation-assessment/summary:
    get:
      summary: 発音分析のサマリー取得
      description: ユーザーの発音分析のサマリーを取得
      parameters:
        - in: header
          name: x-user-id
          schema:
            type: string
          required: true
          description: ユーザーID
      responses:
        '200':
          description: 発音分析のサマリー
          content:
            application/json:
              schema:
                type: object
                properties:
                  overallScore:
                    type: number
                  accuracyScore:
                    type: number
                  fluencyScore:
                    type: number
                  prosodyScore:
                    type: number
                  findings:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum: [ERROR, HABIT]
                        phoneme:
                          type: string
                        details:
                          type: object
        '401':
          description: ユーザーID未指定
        '500':
          description: サーバーエラー

  /pronunciation-assessment/save-raw:
    post:
      summary: 発音評価の生データ保存（DynamoDB）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, unitId, phrase, rawResult]
              properties:
                userId:
                  type: string
                unitId:
                  type: string
                phrase:
                  type: string
                rawResult:
                  type: object
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: 必須項目未指定
        '500':
          description: サーバーエラー

  /pronunciation-assessment/save-summary:
    post:
      summary: 発音の癖集計保存（RDB）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, azureResponse]
              properties:
                userId:
                  type: string
                azureResponse:
                  type: object
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: 必須項目未指定
        '500':
          description: サーバーエラー

  /unit-progress:
    post:
      summary: ユニット進捗保存（UPSERT）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, unitId, status]
              properties:
                userId:
                  type: string
                unitId:
                  type: string
                status:
                  type: string
                lastAttempted:
                  type: string
                  format: date-time
                  nullable: true
                completedAt:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        '200':
          description: 保存成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: 必須項目未指定
        '500':
          description: サーバーエラー
    get:
      summary: ユニット進捗取得
      parameters:
        - in: query
          name: userId
          schema:
            type: string
          required: true
          description: ユーザーID
      responses:
        '200':
          description: 進捗一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
        '400':
          description: ユーザーID未指定
        '500':
          description: サーバーエラー 